Update FermiDiracOneHalfTeSCA
Replace evaluations of log(1+x) with log1p(x).
In the code this occurs twice.

In the old version, trickery in the first branch catched the case
when 1+x == 1 with a test if w==0.

We had this discussion before, the trick came from
https://github.com/JuliaDiff/ForwardDiff.jl/pull/481

This does not work anymore with
https://github.com/JuliaDiff/ForwardDiff.jl/pull/481

This PR (unfortunately) is now in v0.10.33 of ForwardDiff.jl.
There is a discussion going on that this should have been 0.11, i.e. marked
as a breaking change.

It did break Example 103, the PR fixes this.